Worked on Unit Tests for Admin

+++ modules/admin.py
@@ -14,10 +14,6 @@ collection = db["users"]
 
 class Admin(User):
 
-    def __init__(self):
-        pass 
-
-
     def createUser(self, username, type, credit=0):
         """
         Create a User object 
@@ -43,9 +39,10 @@ class Admin(User):
                     #TODOOOOO
                     
                     #add this transaction to the daily transaction file 
-                    transaction = "01" + str(self.username + ("_" * (15 - len(self.username)))) + "_" + self.type + "_" + str(str(self.credit) + ("_" * (9 - len(str(self.credit)))))
+                    transaction = "01 " + str(self.username) + " " + self.type + " " + str(str(self.credit))+"\n"
                     f = open("daily_transaction_file.txt", "a") 
-                    f.write(transaction) 
+                    f.write(transaction)
+                    f.close()
                     
 
                 else:
@@ -117,9 +114,10 @@ class Admin(User):
                 }}
 
                 eventCollection.update_one(eventQuery, remainingTick)
-                transaction = "04" + str(self.username + ("_" * (15 - len(self.username)))) + "_" + str(title + ("_" * (19 - len(title)))) + "_" + ("0" * (3 - len(str(numTickets))) + str(str(numTickets))) + "_" + str(("0" * (6 - len(str(titlePrice)))) + str(titlePrice))
+                transaction = "04 " + str(self.username + (" " * (15 - len(self.username)))) + " " + str(title + (" " * (19 - len(title)))) + " " + ("0" * (3 - len(str(numTickets))) + str(str(numTickets))) + " " + str(("0" * (6 - len(str(titlePrice)))) + str(titlePrice)) +"\n"
                 f = open("daily_transaction_file.txt", "a") 
                 f.write(transaction) 
+                f.close()
                 print("Transaction Confirmed")
                 doStuff = 0
             else:
@@ -137,23 +135,29 @@ class Admin(User):
            raise ValueError("Event cannot have more than 100 tickets")
         #do stuff
         #add to transaction file NOTE: since the event cant sell tickets until after the seller user logs off i think it might be best if we run a routine right before logging out that then adds the event
-        transaction = "04" + str(self.username + ("_" * (15 - len(self.username)))) + "_" + str(title + ("_" * (19 - len(title)))) + "_" + ("0" * (3 - len(str(numTickets))) + str(str(numTickets))) + "_" + str(("0" * (6 - len(str(titlePrice)))) + str(titlePrice))
+        transaction = "04 " + str(self.username + (" " * (15 - len(self.username)))) + " " + str(title + (" " * (19 - len(title)))) + " " + ("0" * (3 - len(str(numTickets))) + str(str(numTickets))) + " " + str(("0" * (6 - len(str(titlePrice)))) + str(titlePrice)) +"\n"
         f = open("daily_transaction_file.txt", "a") 
         f.write(transaction) 
+        f.close()
         print("Event Created - " +"Event Name: " +title +"Ticket Price: " +price +" Number of tickets to be sold: " +numTickets)
 
-    def deleteUser(username):
+    def deleteUser(self, username):
         """
         Deletes user from database
         """
+        if(username == User.getUsername()):
+            raise ValueError("Cannot delete, logged in as user")
+        elif(getUser(username)):
+            #delete the user from the database 
+            collection.delete_one({"username": username})
 
-        #delete the user from the database 
-        collection.delete_one({"username": username})
-
-        #add this transaction to the daily transaction file 
-        transaction = "02" + str(self.username + ("_" * (15 - len(self.username)))) + "_" + self.type + "_" + str(str(self.credit) + ("_" * (9 - len(str(self.credit)))))
-        f = open("daily_transaction_file.txt", "a") 
-        f.write(transaction) 
+            #add this transaction to the daily transaction file 
+            transaction = "02 " + str(self.username) + " " + self.type + " " + str(str(self.credit)) +"\n"
+            f = open("daily_transaction_file.txt", "a") 
+            f.write(transaction)
+            f.close()
+        else:
+            raise ValueError("User not found")
 
     def deleteTicket(ticket):
         """
@@ -201,9 +205,10 @@ class Admin(User):
         """
 
         user = self.getUser(username)
-        
 
         if(credit >= 0):
+            if(credit > 1000):
+                raise ValueError("Value muse not be more than $1000")
             if(user.getCredit() + credit <= 999999):
                 #update credit in database
                 balance = user.getCredit() + credit 
@@ -214,9 +219,10 @@ class Admin(User):
                 collection.update_one(query, newCredit)
 
                 #add the transaction to the daily transaction file 
-                transaction = '06' + "_" + str(self.username + ("_" * (15 - len(self.username)))) + "_" + self.type + "_" + str(str(self.credit) + ("_" * (9 - len(str(self.credit)))))
+                transaction = '06 ' + " " + str(self.username) + " " + self.type + " " + str(str(self.credit))+"\n"
                 f = open("daily_transaction_file.txt", "a") 
-                f.write(transaction) 
+                f.write(transaction)
+                f.close()
             else:
                 raise ValueError("Exceeds credit limit")
         else:
